# loja-virtual/swagger2.json
{
  "swagger": "2.0",
  "info": {
    "description": "Realiza suas compras em nossa loja virtual!",
    "version": "1.0.0",
    "title": "Loja Virtual",
    "contact": {}
  },
  "host": "localhost:8090",
  "basePath": "/api/loja/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/ranking": {
      "get": {
        "tags": [
          "Rankings"
        ],
        "summary": "Operação utilizada para consultar os rankings do produto informado.",
        "description": "Operação utilizada para consultar os rankings do produto informado.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "order",
            "in": "query",
            "description": "Recebe o(s) nome(s) do(s) campo(s) do ranking do produto  a ser ordenado. Campos disponível são: nota",
            "required": false,
            "type": "integer",
            "format": "int32",
            "x-example": "order=nota"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página.",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Total de registros de uma página.",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "idProduto",
            "in": "query",
            "description": "Identificador do produto.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "$ref": "#/definitions/Rankings"
            }
          },
          "400": {
            "description": "Requisição mal formada."
          },
          "401": {
            "description": "Requisição requer autenticação."
          },
          "403": {
            "description": "Requisição negada."
          }
        }
      },
      "post": {
        "tags": [
          "Rankings"
        ],
        "summary": "Operação utilizada para criar um ranking do produto.",
        "description": "Operação utilizada para criar um ranking do produto.",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Ranking"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Criado com sucesso."
          },
          "400": {
            "description": "Requisição mal formada."
          },
          "401": {
            "description": "Requisição requer autenticação."
          },
          "403": {
            "description": "Requisição negada."
          }
        }
      }
    },
    "/ranking/notas/{idProduto}": {
      "get": {
        "tags": [
          "Rankings"
        ],
        "summary": "Operação utilizada para calcular as notas do ranking do produto.",
        "description": "Operação utilizada para calcular as notas do ranking do produto.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idProduto",
            "in": "path",
            "description": "Identificador do produto.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "$ref": "#/definitions/Nota"
            }
          },
          "400": {
            "description": "Requisição mal formada."
          },
          "401": {
            "description": "Requisição requer autenticação."
          },
          "403": {
            "description": "Requisição negada."
          }
        }
      }
    },
    "/ranking/{idRanking}": {
      "get": {
        "tags": [
          "Rankings"
        ],
        "summary": "Operação utilizada para se obter um ranking.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idRanking",
            "in": "path",
            "description": "Identificador do Ranking.",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "schema": {
              "$ref": "#/definitions/Ranking"
            }
          },
          "400": {
            "description": "Requisição mal formada."
          },
          "401": {
            "description": "Requisição requer autenticação."
          },
          "403": {
            "description": "Requisição negada."
          },
          "404": {
            "description": "Recurso não encontrado."
          }
        }
      }
    }
  },
  "definitions": {
...
    "Metadata": {
      "type": "object",
      "properties": {
        "chave": {
          "type": "string",
          "description": "Chave."
        },
        "valor": {
          "type": "string",
          "description": "Valor."
        }
      },
      "description": "Representa um par de chave/valor."
    },
    "Nota": {
      "type": "object",
      "properties": {
        "maiorNota": {
          "type": "integer",
          "format": "int64",
          "description": "Maior nota do ranking."
        },
        "media": {
          "type": "integer",
          "format": "int64",
          "description": "Média do ranking."
        },
        "menorNota": {
          "type": "integer",
          "format": "int64",
          "description": "Menor nota do ranking."
        }
      },
      "description": "Representa uma nota."
    },
...
    "Ranking": {
      "type": "object",
      "required": [
        "idProduto",
        "nota"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "Identificador do ranking."
        },
        "idProduto": {
          "type": "integer",
          "format": "int64",
          "description": "Produto do ranking."
        },
        "nota": {
          "type": "integer",
          "format": "int32",
          "description": "Nota do ranking."
        },
        "comentario": {
          "type": "string",
          "description": "Comentário do ranking."
        }
      },
      "description": "Representa o ranking de um produto."
    },
    "Rankings": {
      "type": "object",
      "properties": {
        "rankings": {
          "type": "array",
          "description": "Lista de rankings do produto.",
          "items": {
            "$ref": "#/definitions/Ranking"
          }
        },
        "metadata": {
          "type": "array",
          "description": "Informações de paginação.",
          "items": {
            "$ref": "#/definitions/Metadata"
          }
        }
      },
      "description": "Encapsula uma lista de rankings do produto e meta informações, tais como paginação."
    }
  }
}